+++ R/spec_rjd.R
@@ -17,7 +17,7 @@ spec_regarima_X13_jd2r <- function(spec = NA, context_dictionary = NULL,
                          freq = NA){
 
   #Estimate
-  preliminary.check <- spec$getBasic()$isPreliminaryCheck()
+  preliminary.check <- .jcall(.jcall(spec, "Ljdr/spec/x13/BasicSpec;" ,"getBasic"), "Z", "isPreliminaryCheck")
 
   jestimate <-.jcall(spec,"Ljdr/spec/x13/EstimateSpec;","getEstimate")
   jest.span <-.jcall(jestimate,"Ljdr/spec/ts/SpanSelector;","getSpan")
@@ -173,9 +173,9 @@ spec_regarima_X13_jd2r <- function(spec = NA, context_dictionary = NULL,
              "getPrespecifiedOutlier",
              as.integer(i-1))
     })
-    type <- sapply(outliers, function(x) x$getCode())
-    date <- sapply(outliers, function(x) x$getPosition())
-    coeff <- sapply(outliers, function(x) x$getCoefficient())
+    type <- sapply(outliers, .jcall, "S", "getCode")
+    date <- sapply(outliers, .jcall, "S", "getPosition")
+    coeff <- sapply(outliers, .jcall, "D", "getCoefficient")
     if(all(coeff == 0)){ #All coefficients are equal to 0: they are not fixed
       result$userdef_spec$specification$outlier.coef <- FALSE
       coeff <- coeff * NA
@@ -200,10 +200,26 @@ spec_regarima_X13_jd2r <- function(spec = NA, context_dictionary = NULL,
              as.integer(i-1))
     })
 
-    type <- sapply(ud_vars, function(x) x$getComponent())
-    coeff <- sapply(ud_vars, function(x) x$getCoefficient())
-    var_names <- sapply(ud_vars, function(x) x$getName())
+    type <- sapply(ud_vars, .jcall, "S", "getComponent")
+    coeff <- sapply(ud_vars, .jcall, "D", "getCoefficient")
+    var_names <- sapply(ud_vars, .jcall, "S", "getName")
+
+
     var_names_split <- strsplit(var_names,"[.]")
+    core_regression <- jregression$getCore()$getRegression()
+    if (core_regression$hasFixedCoefficients()) {
+      coeff <- sapply(seq_along(var_names_split), function(i){
+        td_name <- paste0(var_names_split[[i]], collapse = "@")
+        print(td_name)
+        fixed_coeff <- .jcall(core_regression, "[D", "getFixedCoefficients", td_name)
+        if (is.null(fixed_coeff)) {
+          0
+        }else{
+          fixed_coeff
+        }
+      })
+      result$userdef_spec$specification$variables.coef <- TRUE
+    }
     var_names <- sapply(var_names_split, function(x) x[2])
     var_names <- base::make.names(var_names, unique = TRUE)
     var_names <- gsub(".","_", var_names, fixed = TRUE)
@@ -226,9 +242,12 @@ spec_regarima_X13_jd2r <- function(spec = NA, context_dictionary = NULL,
     if(!is.null(context_dictionary)){
 
       var_series <- lapply(var_names_split,function(names){
-        ts_variable <- context_dictionary$getTsVariable(names[1],
-                                                         names[2])
-        ts_jd2r(ts_variable$getTsData())
+        ts_variable <- .jcall(context_dictionary,
+                              "Lec/tstoolkit/timeseries/regression/ITsVariable;",
+                              "getTsVariable",
+                              names[1],
+                              names[2])
+        ts_jd2r(.jcall(ts_variable, "Lec/tstoolkit/timeseries/simplets/TsData;", "getTsData"))
       })
       var_series <- ts(simplify2array(var_series),
                        start = start(var_series[[1]]), frequency = frequency(var_series[[1]]))
@@ -242,7 +261,7 @@ spec_regarima_X13_jd2r <- function(spec = NA, context_dictionary = NULL,
   }
 
   #Calendar
-  user_td <- jtd$getUserVariables()
+  user_td <- .jcall(jtd, "[S", "getUserVariables")
   if(length(user_td) > 0 ){
     var_names_split <- strsplit(user_td,"[.]")
     var_names <- sapply(var_names_split, function(x) x[2])
@@ -251,9 +270,23 @@ spec_regarima_X13_jd2r <- function(spec = NA, context_dictionary = NULL,
 
     result$userdef_spec$specification$variables <-
       TRUE
+    coeff <- NA
+    core_regression <- jregression$getCore()$getRegression()
+    if (core_regression$hasFixedCoefficients()) {
+      coeff <- sapply(seq_along(var_names_split), function(i){
+        td_name <- paste0(var_names_split[[i]], collapse = "@")
+        fixed_coeff <- .jcall(core_regression, "[D", "getFixedCoefficients", td_name)
+        if (is.null(fixed_coeff)) {
+          NA
+        }else{
+          fixed_coeff
+        }
+      })
+      result$userdef_spec$specification$variables.coef <- TRUE
+    }
 
     td_var_description <- data.frame(type = rep("Calendar",length(var_names)),
-                                     coeff = NA, row.names = var_names)
+                                     coeff = coeff, row.names = var_names)
     if(identical_na(result$userdef_spec$variables$description)){
       result$userdef_spec$variables$description <- td_var_description
     }else{
@@ -264,9 +297,12 @@ spec_regarima_X13_jd2r <- function(spec = NA, context_dictionary = NULL,
     if(!is.null(context_dictionary)){
 
       var_series <- lapply(var_names_split,function(names){
-        ts_variable <- context_dictionary$getTsVariable(names[1],
-                                                         names[2])
-        ts_jd2r(ts_variable$getTsData())
+        ts_variable <- .jcall(context_dictionary,
+                              "Lec/tstoolkit/timeseries/regression/ITsVariable;",
+                              "getTsVariable",
+                              gsub("^td\\|", "", names[1]),
+                              names[2])
+        ts_jd2r(.jcall(ts_variable, "Lec/tstoolkit/timeseries/simplets/TsData;", "getTsData"))
       })
       var_series <- ts(simplify2array(var_series),
                        start = start(var_series[[1]]), frequency = frequency(var_series[[1]]))
@@ -282,20 +318,20 @@ spec_regarima_X13_jd2r <- function(spec = NA, context_dictionary = NULL,
 
   ## Ramp effects
   core_regression <- jregression$getCore()$getRegression()
-  jramps <- core_regression$getRamps()
-  nb_ramps <- core_regression$getRampsCount()
+  jramps <- .jcall(core_regression, "[Lec/tstoolkit/timeseries/regression/Ramp;", "getRamps")
+  nb_ramps <- .jcall(core_regression, "I", "getRampsCount")
   if (nb_ramps > 0) {
-    var_names <- sapply(seq_len(nb_ramps), function(x) jramps[[x]]$getDescription())
+    var_names <- sapply(jramps, .jcall, "S", "getDescription")
 
     result$userdef_spec$specification$variables <-
       TRUE
 
     coeff <- NA
-    if (core_regression$hasFixedCoefficients()) {
+    if (.jcall(core_regression, "Z", "hasFixedCoefficients")) {
       coeff <- sapply(seq_len(nb_ramps), function(i){
         jramp <- jramps[[i]]
-        ramp_name <- jramp$getName()
-        fixed_coeff <- core_regression$getFixedCoefficients(ramp_name)
+        ramp_name <- .jcall(jramp, "S", "getName")
+        fixed_coeff <- .jcall(core_regression, "[D", "getFixedCoefficients", ramp_name)
         if (is.null(fixed_coeff)) {
           NA
         }else{
@@ -325,11 +361,13 @@ spec_regarima_X13_jd2r <- function(spec = NA, context_dictionary = NULL,
         end_ts <- end(result$userdef_spec$variables$series)
         ramp_series <- lapply(seq_len(nb_ramps), function(i){
           jramp <- jramps[[i]]
-          jstart <- jramp$getStart()
-          jend <- jramp$getEnd()
+          jstart <- .jcall(jramp, "Lec/tstoolkit/timeseries/Day;", "getStart")
+          jend <- .jcall(jramp, "Lec/tstoolkit/timeseries/Day;", "getEnd")
 
-          start_ramp <- c(jstart$getYear(), jstart$getMonth())
-          end_ramp <- c(jend$getYear(), jend$getMonth())
+          start_ramp <- c(.jcall(jstart, "I", "getYear"),
+                          .jcall(jstart, "I", "getMonth"))
+          end_ramp <- c(.jcall(jend, "I", "getYear"),
+                        .jcall(jend, "I", "getMonth"))
           ramp(start = start_ts, end = end_ts,
                start_ramp = start_ramp, end_ramp = end_ramp,
                frequency = freq)
@@ -337,11 +375,13 @@ spec_regarima_X13_jd2r <- function(spec = NA, context_dictionary = NULL,
       } else {
         ramp_series <- lapply(seq_len(nb_ramps), function(i){
           jramp <- jramps[[i]]
-          jstart <- jramp$getStart()
-          jend <- jramp$getEnd()
+          jstart <- .jcall(jramp, "Lec/tstoolkit/timeseries/Day;", "getStart")
+          jend <- .jcall(jramp, "Lec/tstoolkit/timeseries/Day;", "getEnd")
 
-          start_ramp <- c(jstart$getYear(), jstart$getMonth())
-          end_ramp <- c(jend$getYear(), jend$getMonth())
+          start_ramp <- c(.jcall(jstart, "I", "getYear"),
+                          .jcall(jstart, "I", "getMonth"))
+          end_ramp <- c(.jcall(jend, "I", "getYear"),
+                        .jcall(jend, "I", "getMonth"))
           ramp(start_ramp = start_ramp, end_ramp = end_ramp,
                frequency = freq)
         })
@@ -350,7 +390,13 @@ spec_regarima_X13_jd2r <- function(spec = NA, context_dictionary = NULL,
                         start = start(ramp_series[[1]]),
                         frequency = frequency(ramp_series[[1]]))
       if (!identical_na(result$userdef_spec$variables$series)) {
-        ramp_series <- ts.union(result$userdef_spec$variables$series, ramp_series)
+        if (frequency (result$userdef_spec$variables$series) != frequency(ramp_series)){
+          # here we assume that the seris in result$userdef_spec$variables$series are wrong and are then deleted
+          # e.g calendar regressors defined in spec monthly while the input ts is quarterly
+          ramp_series <- ramp_series
+        } else {
+          ramp_series <- ts.union(result$userdef_spec$variables$series, ramp_series)
+        }
       }
       if (is.mts(ramp_series))
         colnames(ramp_series) <- rownames(result$userdef_spec$variables$description)
@@ -359,10 +405,10 @@ spec_regarima_X13_jd2r <- function(spec = NA, context_dictionary = NULL,
     }
   }
 
-  Phi <- jarima$getPhi()
-  BPhi <- jarima$getBPhi()
-  Theta <- jarima$getTheta()
-  BTheta <- jarima$getBTheta()
+  Phi <- .jcall(jarima,"[Lec/tstoolkit/Parameter;","getPhi")
+  BPhi <- .jcall(jarima,"[Lec/tstoolkit/Parameter;","getBPhi")
+  Theta <- .jcall(jarima,"[Lec/tstoolkit/Parameter;","getTheta")
+  BTheta <- .jcall(jarima,"[Lec/tstoolkit/Parameter;","getBTheta")
   arima_coefficients_spec <-
     rbind(arimaCoef_jd2r(Phi),
           arimaCoef_jd2r(BPhi),
@@ -382,7 +428,7 @@ spec_TRAMO_jd2r <- function(spec = NA, context_dictionary = NULL,
                        extra_info = FALSE, freq = NA){
 
   #Estimate
-  preliminary.check <- spec$getBasic()$isPreliminaryCheck()
+  preliminary.check <- .jcall(.jcall(spec, "Ljdr/spec/tramoseats/BasicSpec;" ,"getBasic"), "Z", "isPreliminaryCheck")
   jestimate <- .jcall(spec,"Ljdr/spec/tramoseats/EstimateSpec;","getEstimate")
   jest.span <- .jcall(jestimate,"Ljdr/spec/ts/SpanSelector;","getSpan")
 
@@ -534,9 +580,9 @@ spec_TRAMO_jd2r <- function(spec = NA, context_dictionary = NULL,
              "getPrespecifiedOutlier",
              as.integer(i-1))
     })
-    type <- sapply(outliers, function(x) x$getCode())
-    date <- sapply(outliers, function(x) x$getPosition())
-    coeff <- sapply(outliers, function(x) x$getCoefficient())
+    type <- sapply(outliers, .jcall, "S", "getCode")
+    date <- sapply(outliers, .jcall, "S", "getPosition")
+    coeff <- sapply(outliers, .jcall, "D", "getCoefficient")
 
     if(all(coeff == 0)){ #All coefficients are equal to 0: they are not fixed
       result$userdef_spec$specification$outlier.coef <- FALSE
@@ -564,12 +610,25 @@ spec_TRAMO_jd2r <- function(spec = NA, context_dictionary = NULL,
              as.integer(i-1))
     })
 
-    type <- sapply(ud_vars, function(x) x$getComponent())
-    coeff <- sapply(ud_vars, function(x) x$getCoefficient())
-    var_names <- sapply(ud_vars, function(x) x$getName())
+    type <- sapply(ud_vars, .jcall, "S", "getComponent")
+    coeff <- sapply(ud_vars, .jcall, "D", "getCoefficient")
+    var_names <- sapply(ud_vars, .jcall, "S", "getName")
     var_names_split <- strsplit(var_names,"[.]")
     var_names <- sapply(var_names_split, function(x) x[2])
 
+    core_regression <- jregression$getCore()$getRegression()
+    if (core_regression$hasFixedCoefficients()) {
+      coeff <- sapply(seq_along(var_names_split), function(i){
+        td_name <- paste0(var_names_split[[i]], collapse = "@")
+        fixed_coeff <- .jcall(core_regression, "[D", "getFixedCoefficients", td_name)
+        if (is.null(fixed_coeff)) {
+          0
+        }else{
+          fixed_coeff
+        }
+      })
+      result$userdef_spec$specification$variables.coef <- TRUE
+    }
     if(all(coeff == 0)){ #All coefficients are equal to 0: they are not fixed
       result$userdef_spec$specification$variables.coef <- FALSE
       coeff <- coeff * NA
@@ -588,9 +647,12 @@ spec_TRAMO_jd2r <- function(spec = NA, context_dictionary = NULL,
     if(!is.null(context_dictionary)){
 
       var_series <- lapply(var_names_split,function(names){
-        ts_variable <- context_dictionary$getTsVariable(names[1],
-                                                         names[2])
-        ts_jd2r(ts_variable$getTsData())
+        ts_variable <- .jcall(context_dictionary,
+                              "Lec/tstoolkit/timeseries/regression/ITsVariable;",
+                              "getTsVariable",
+                              names[1],
+                              names[2])
+        ts_jd2r(.jcall(ts_variable, "Lec/tstoolkit/timeseries/simplets/TsData;", "getTsData"))
       })
       var_series <- ts(simplify2array(var_series),
                        start = start(var_series[[1]]), frequency = frequency(var_series[[1]]))
@@ -604,7 +666,7 @@ spec_TRAMO_jd2r <- function(spec = NA, context_dictionary = NULL,
   }
 
   #Calendar
-  user_td <- jtd$getUserVariables()
+  user_td <- .jcall(jtd, "[S", "getUserVariables")
   if(length(user_td) > 0 ){
     var_names_split <- strsplit(user_td,"[.]")
     var_names <- sapply(var_names_split, function(x) x[2])
@@ -612,19 +674,19 @@ spec_TRAMO_jd2r <- function(spec = NA, context_dictionary = NULL,
     result$userdef_spec$specification$variables <-
       TRUE
     coeff <- NA
-    # if (core_regression$hasFixedCoefficients()) {
-    #   # coeff <- sapply(seq_len(nb_ramps), function(i){
-    #   #   jramp <- jramps[[i]]
-    #   #   ramp_name <- jramp$getName()
-    #   #   fixed_coeff <- core_regression$getFixedCoefficients(ramp_name)
-    #   #   if (is.null(fixed_coeff)) {
-    #   #     NA
-    #   #   }else{
-    #   #     fixed_coeff
-    #   #   }
-    #   # })
-    #   # result$userdef_spec$specification$variables.coef <- TRUE
-    # }
+    core_regression <- jregression$getCore()$getRegression()
+    if (core_regression$hasFixedCoefficients()) {
+      coeff <- sapply(seq_along(var_names_split), function(i){
+        td_name <- paste0(var_names_split[[i]], collapse = "@")
+        fixed_coeff <- .jcall(core_regression, "[D", "getFixedCoefficients", td_name)
+        if (is.null(fixed_coeff)) {
+          NA
+        }else{
+          fixed_coeff
+        }
+      })
+      result$userdef_spec$specification$variables.coef <- TRUE
+    }
 
     td_var_description <- data.frame(type = rep("Calendar",length(var_names)),
                                      coeff = coeff, row.names = var_names)
@@ -638,9 +700,12 @@ spec_TRAMO_jd2r <- function(spec = NA, context_dictionary = NULL,
     if(!is.null(context_dictionary)){
 
       var_series <- lapply(var_names_split,function(names){
-        ts_variable <- context_dictionary$getTsVariable(names[1],
-                                                         names[2])
-        ts_jd2r(ts_variable$getTsData())
+        ts_variable <- .jcall(context_dictionary,
+                              "Lec/tstoolkit/timeseries/regression/ITsVariable;",
+                              "getTsVariable",
+                              gsub("^td\\|", "", names[1]),
+                              names[2])
+        ts_jd2r(.jcall(ts_variable, "Lec/tstoolkit/timeseries/simplets/TsData;", "getTsData"))
       })
       var_series <- ts(simplify2array(var_series),
                        start = start(var_series[[1]]), frequency = frequency(var_series[[1]]))
@@ -656,19 +721,19 @@ spec_TRAMO_jd2r <- function(spec = NA, context_dictionary = NULL,
 
   ## Ramp effects
   core_regression <- jregression$getCore()$getRegression()
-  jramps <- core_regression$getRamps()
-  nb_ramps <- core_regression$getRampsCount()
+  jramps <- .jcall(core_regression, "[Lec/tstoolkit/timeseries/regression/Ramp;", "getRamps")
+  nb_ramps <- .jcall(core_regression, "I", "getRampsCount")
   if (nb_ramps > 0) {
-    var_names <- sapply(seq_len(nb_ramps), function(x) jramps[[x]]$getDescription())
+    var_names <- sapply(jramps, .jcall, "S", "getDescription")
 
     result$userdef_spec$specification$variables <-
       TRUE
     coeff <- NA
-    if (core_regression$hasFixedCoefficients()) {
+    if (.jcall(core_regression, "Z", "hasFixedCoefficients")) {
       coeff <- sapply(seq_len(nb_ramps), function(i){
         jramp <- jramps[[i]]
-        ramp_name <- jramp$getName()
-        fixed_coeff <- core_regression$getFixedCoefficients(ramp_name)
+        ramp_name <- .jcall(jramp, "S", "getName")
+        fixed_coeff <- .jcall(core_regression, "[D", "getFixedCoefficients", ramp_name)
         if (is.null(fixed_coeff)) {
           NA
         }else{
@@ -698,11 +763,13 @@ spec_TRAMO_jd2r <- function(spec = NA, context_dictionary = NULL,
         end_ts <- end(result$userdef_spec$variables$series)
         ramp_series <- lapply(seq_len(nb_ramps), function(i){
           jramp <- jramps[[i]]
-          jstart <- jramp$getStart()
-          jend <- jramp$getEnd()
+          jstart <- .jcall(jramp, "Lec/tstoolkit/timeseries/Day;", "getStart")
+          jend <- .jcall(jramp, "Lec/tstoolkit/timeseries/Day;", "getEnd")
 
-          start_ramp <- c(jstart$getYear(), jstart$getMonth())
-          end_ramp <- c(jend$getYear(), jend$getMonth())
+          start_ramp <- c(.jcall(jstart, "I", "getYear"),
+                          .jcall(jstart, "I", "getMonth"))
+          end_ramp <- c(.jcall(jend, "I", "getYear"),
+                        .jcall(jend, "I", "getMonth"))
           ramp(start = start_ts, end = end_ts,
                start_ramp = start_ramp, end_ramp = end_ramp,
                frequency = freq)
@@ -710,11 +777,13 @@ spec_TRAMO_jd2r <- function(spec = NA, context_dictionary = NULL,
       } else {
         ramp_series <- lapply(seq_len(nb_ramps), function(i){
           jramp <- jramps[[i]]
-          jstart <- jramp$getStart()
-          jend <- jramp$getEnd()
+          jstart <- .jcall(jramp, "Lec/tstoolkit/timeseries/Day;", "getStart")
+          jend <- .jcall(jramp, "Lec/tstoolkit/timeseries/Day;", "getEnd")
 
-          start_ramp <- c(jstart$getYear(), jstart$getMonth())
-          end_ramp <- c(jend$getYear(), jend$getMonth())
+          start_ramp <- c(.jcall(jstart, "I", "getYear"),
+                          .jcall(jstart, "I", "getMonth"))
+          end_ramp <- c(.jcall(jend, "I", "getYear"),
+                        .jcall(jend, "I", "getMonth"))
           ramp(start_ramp = start_ramp, end_ramp = end_ramp,
                frequency = freq)
         })
@@ -723,7 +792,13 @@ spec_TRAMO_jd2r <- function(spec = NA, context_dictionary = NULL,
                        start = start(ramp_series[[1]]),
                        frequency = frequency(ramp_series[[1]]))
       if (!identical_na(result$userdef_spec$variables$series)) {
-        ramp_series <- ts.union(result$userdef_spec$variables$series, ramp_series)
+        if (frequency (result$userdef_spec$variables$series) != frequency(ramp_series)){
+          # here we assume that the series in result$userdef_spec$variables$series are wrong and are then deleted
+          # e.g calendar regressors defined in spec monthly while the input ts is quarterly
+          ramp_series <- ramp_series
+        } else {
+          ramp_series <- ts.union(result$userdef_spec$variables$series, ramp_series)
+        }
       }
       if (is.mts(ramp_series))
         colnames(ramp_series) <- rownames(result$userdef_spec$variables$description)
@@ -733,10 +808,10 @@ spec_TRAMO_jd2r <- function(spec = NA, context_dictionary = NULL,
   }
 
   #Arima
-  Phi <- jarima$getPhi()
-  BPhi <- jarima$getBPhi()
-  Theta <- jarima$getTheta()
-  BTheta <- jarima$getBTheta()
+  Phi <- .jcall(jarima,"[Lec/tstoolkit/Parameter;","getPhi")
+  BPhi <- .jcall(jarima,"[Lec/tstoolkit/Parameter;","getBPhi")
+  Theta <- .jcall(jarima,"[Lec/tstoolkit/Parameter;","getTheta")
+  BTheta <- .jcall(jarima,"[Lec/tstoolkit/Parameter;","getBTheta")
   arima_coefficients_spec <-
     rbind(arimaCoef_jd2r(Phi),
           arimaCoef_jd2r(BPhi),
@@ -776,7 +851,9 @@ specX11_jd2r <- function(spec = NA, freq = NA){
   fcasts <- .jcall(jx11,"I","getForecastHorizon")
   bcasts <- .jcall(jx11,"I","getBackcastHorizon")
   excludeFcasts <- .jcall(jx11,"Z","isExcludefcst")
-  calendarSigma <- jx11$getCalendarSigma()$toString()
+  calendarSigma <- .jcall(
+    .jcall(jx11, "Lec/satoolkit/x11/CalendarSigma;", "getCalendarSigma"),
+    "S", "toString")
   sigmaVector <- .jcall(jx11,returnSig = "[S", "getSigmavec") #TODO
   if (length(sigmaVector) == 0) {
     sigmaVector <- NA
@@ -867,8 +944,9 @@ preVar_r2jd <- function(jsobjct = NA, jsdict = NA, coefEna = NA,
     if(n_calendar_def == 1){
       .jcall(jsdict,"V","add",varNames,ts_r2jd(series))
       if(coef!=0){
-        jcoreg$setFixedCoefficients(paste0("td|r@",varNames),
-                                    .jarray(coef))
+        .jcall(jcoreg, "V", "setFixedCoefficients",
+               paste0("td|r@",varNames),
+               .jarray(coef))
       }
       .jcall(jtd,"V","setUserVariables", .jarray(paste0("r.",varNames)))
 
@@ -884,8 +962,9 @@ preVar_r2jd <- function(jsobjct = NA, jsdict = NA, coefEna = NA,
         for (i in 1:nvar){
           .jcall(jsdict,"V","add",varNames[i],ts_r2jd(series[,i]))
           if(coef[i]!=0){
-            jcoreg$setFixedCoefficients(paste0("td|r@",varNames[i]),
-                                        .jarray(coef[i]))
+            .jcall(jcoreg, "V", "setFixedCoefficients",
+                   paste0("td|r@",varNames[i]),
+                   .jarray(coef[i]))
           }
         }
         .jcall(jtd,"V","setUserVariables",
@@ -911,8 +990,9 @@ preVar_r2jd <- function(jsobjct = NA, jsdict = NA, coefEna = NA,
         .jcall(jsdict,"V","add",varNames[i],
                ts_r2jd(series[, i]))
         if(coef[i]!=0){
-          jcoreg$setFixedCoefficients(paste0("td|r@",varNames[i]),
-                                      .jarray(coef[i]))
+          .jcall(jcoreg, "V", "setFixedCoefficients",
+                 paste0("td|r@",varNames[i]),
+                 .jarray(coef[i]))
         }
       }
       .jcall(jtd,"V","setUserVariables",
@@ -960,8 +1040,8 @@ arimaCoef_jd2r <- function(jparams){
   if (len==0)
     return (NULL)
   param_name <- deparse(substitute(jparams))
-  Type <- sapply(param, function(x) x$getType()$toString())
-  Value <- sapply(param, function(x) x$getValue())
+  Type <- sapply(param, function(x) .jcall(.jcall(x, "Lec/tstoolkit/ParameterType;" , "getType"), "S", "toString"))
+  Value <- sapply(param, .jcall, "D", "getValue")
   data_param <- data.frame(Type = Type, Value = Value)
   rownames(data_param) <- sprintf("%s(%i)",
                                   param_name,
@@ -991,7 +1071,7 @@ spec_regarima_X13_r2jd <- function(spec = NA, jdspec = NA){
             n0 = as.integer(span[1,4]), n1=as.integer(span[1,5]))
   .jcall(jestimate ,"V","setTol", as.numeric(est["tolerance"]))
 
-  jdspec$getBasic()$setPreliminaryCheck(est[1, "preliminary.check"])
+  .jcall(.jcall(jdspec, "Ljdr/spec/x13/BasicSpec;" ,"getBasic"), "V", "setPreliminaryCheck", est[1, "preliminary.check"])
 
   #Transform
   jtransform <-.jcall(jdspec,"Ljdr/spec/x13/TransformSpec;","getTransform")
@@ -1110,7 +1190,7 @@ spec_TRAMO_r2jd <- function(spec = NA, jdspec =NA){
   span <- s_span(spec)
 
   #Estimate
-  jdspec$getBasic()$setPreliminaryCheck(est[1, "preliminary.check"])
+  .jcall(.jcall(jdspec, "Ljdr/spec/tramoseats/BasicSpec;" ,"getBasic"), "V", "setPreliminaryCheck", est[1, "preliminary.check"])
 
   jestimate <-.jcall(jdspec,"Ljdr/spec/tramoseats/EstimateSpec;","getEstimate")