+++ R/tramoseats.R
@@ -150,17 +150,20 @@ tramoseats.SA_spec <- function(series, spec,
   }else{
 
     # Error during the preliminary check
-    res = jrslt$getResults()$getProcessingInformation()
+    proc_info <- .jcall(
+      .jcall(jrslt,
+             "Lec/tstoolkit/algorithm/CompositeResults;", "getResults"
+      ),
+      "Ljava/util/List;", "getProcessingInformation"
+    )
 
-    if(is.null(jrslt$getDiagnostics()) & !.jcall(res,"Z","isEmpty")){
-      proc_info <- jrslt$getResults()$getProcessingInformation()
-
+    if(is.null(.jcall(jrslt,"Lec/tstoolkit/jdr/sa/SaDiagnostics;", "getDiagnostics")) & !.jcall(proc_info,"Z","isEmpty")){
       error_msg <- .jcall(proc_info, "Ljava/lang/Object;", "get", 0L)$getErrorMessages(proc_info)
       warning_msg <- .jcall(proc_info, "Ljava/lang/Object;", "get", 0L)$getWarningMessages(proc_info)
       if(!.jcall(error_msg,"Z","isEmpty"))
-        stop(error_msg$toString())
+        stop(.jcall(error_msg, "S", "toString"))
       if(!.jcall(warning_msg,"Z","isEmpty"))
-        warning(warning_msg$toString())
+        warning(.jcall(warning_msg, "S", "toString"))
     }
     reg <- regarima_TS(jrobj = jrobct_arima, spec = spec$regarima)
     deco <- decomp_TS(jrobj = jrobct, spec = spec$seats)
@@ -199,16 +202,20 @@ tramoseatsJavaResults <- function(jrslt, spec,
     return(NaN)
 
   # Error in preliminary check
-  res = jrslt$getResults()$getProcessingInformation()
+  proc_info <- .jcall(
+    .jcall(jrslt,
+           "Lec/tstoolkit/algorithm/CompositeResults;", "getResults"
+    ),
+    "Ljava/util/List;", "getProcessingInformation"
+  )
 
-  if(is.null(jrslt$getDiagnostics()) & !.jcall(res,"Z","isEmpty")){
-    proc_info <- jrslt$getResults()$getProcessingInformation()
+  if(is.null(.jcall(jrslt,"Lec/tstoolkit/jdr/sa/SaDiagnostics;", "getDiagnostics")) & !.jcall(proc_info,"Z","isEmpty")){
     error_msg <- .jcall(proc_info, "Ljava/lang/Object;", "get", 0L)$getErrorMessages(proc_info)
     warning_msg <- .jcall(proc_info, "Ljava/lang/Object;", "get", 0L)$getWarningMessages(proc_info)
     if(!.jcall(error_msg,"Z","isEmpty"))
-      stop(error_msg$toString())
+      stop(.jcall(error_msg, "S", "toString"))
     if(!.jcall(warning_msg,"Z","isEmpty"))
-      warning(warning_msg$toString())
+      warning(.jcall(warning_msg, "S", "toString"))
   }
 
   reg <- regarima_defTS(jrobj = jrobct_arima, spec = spec,