+++ R/x13.R
@@ -146,16 +146,20 @@ x13.SA_spec <- function(series, spec, userdefined = NULL){
     return(NaN)
   }else{
     # Error with the preliminary check
-    res = jrslt$getResults()$getProcessingInformation()
+    proc_info <- .jcall(
+      .jcall(jrslt,
+             "Lec/tstoolkit/algorithm/CompositeResults;", "getResults"
+      ),
+      "Ljava/util/List;", "getProcessingInformation"
+    )
 
-    if(is.null(jrslt$getDiagnostics()) & !.jcall(res,"Z","isEmpty")){
-      proc_info <- jrslt$getResults()$getProcessingInformation()
+    if(is.null(.jcall(jrslt,"Lec/tstoolkit/jdr/sa/SaDiagnostics;", "getDiagnostics")) & !.jcall(proc_info,"Z","isEmpty")){
       error_msg <- .jcall(proc_info, "Ljava/lang/Object;", "get", 0L)$getErrorMessages(proc_info)
       warning_msg <- .jcall(proc_info, "Ljava/lang/Object;", "get", 0L)$getWarningMessages(proc_info)
       if(!.jcall(error_msg,"Z","isEmpty"))
-        stop(error_msg$toString())
+        stop(.jcall(error_msg, "S", "toString"))
       if(!.jcall(warning_msg,"Z","isEmpty"))
-        warning(warning_msg$toString())
+        warning(.jcall(warning_msg, "S", "toString"))
     }
     reg <- regarima_X13(jrobj = jrobct_arima, spec = spec$regarima)
     deco <- decomp_X13(jrobj = jrobct, spec = spec$x11, seasma = seasma)
@@ -195,15 +199,19 @@ x13JavaResults <- function(jrslt, spec, userdefined = NULL,
   }
 
   # Error during the preliminary check
-  res = jrslt$getResults()$getProcessingInformation()
-  if(is.null(jrslt$getDiagnostics()) & !.jcall(res,"Z","isEmpty")){
-    proc_info <- jrslt$getResults()$getProcessingInformation()
+  proc_info <- .jcall(
+    .jcall(jrslt,
+           "Lec/tstoolkit/algorithm/CompositeResults;", "getResults"
+    ),
+    "Ljava/util/List;", "getProcessingInformation"
+  )
+  if(is.null(.jcall(jrslt,"Lec/tstoolkit/jdr/sa/SaDiagnostics;", "getDiagnostics")) & !.jcall(proc_info,"Z","isEmpty")){
     error_msg <- .jcall(proc_info, "Ljava/lang/Object;", "get", 0L)$getErrorMessages(proc_info)
     warning_msg <- .jcall(proc_info, "Ljava/lang/Object;", "get", 0L)$getWarningMessages(proc_info)
     if(!.jcall(error_msg,"Z","isEmpty"))
-      stop(error_msg$toString())
+      stop(.jcall(error_msg, "S", "toString"))
     if(!.jcall(warning_msg,"Z","isEmpty"))
-      warning(warning_msg$toString())
+      warning(.jcall(warning_msg, "S", "toString"))
   }
 
   reg <- regarima_defX13(jrobj = jrobct_arima, spec = spec,